requirements.txt
Flask>=2.1
Flask-WTF>=1.0
Flask-Login>=0.6
Flask-SQLAlchemy>=3.0
python-dotenv>=1.0
mercadopago>=2.0.0
Werkzeug>=2.1
email-validator>=1.3
